fix: use handle reponse http

+++ src/modules/clients/clients.controller.ts
@@ -3,6 +3,7 @@ import { ClientsService } from './clients.service'
 import { ApiTags } from '@nestjs/swagger'
 import { Post, Body, Get, Param, Delete } from '@nestjs/common'
 import { CreateClientDto } from './dto/client.dto'
+import { handleBadresuest } from 'src/common/handleHttp'
 
 @ApiTags('clients')
 @Controller('clients')
@@ -11,11 +12,16 @@ export class ClientsController {
 
   @Post()
   async createClient(@Body() client: CreateClientDto) {
+    if (!client)
+      return handleBadresuest(
+        new Error('Porfavor envie un cliente que desea registrar'),
+      )
     return await this.clientsService.createClient(client)
   }
 
   @Get(':id')
   async findById(@Param('id') id: number) {
+    if (!id) return handleBadresuest(new Error('El id es requerido'))
     return await this.clientsService.findById(id)
   }
 
@@ -26,6 +32,7 @@ export class ClientsController {
 
   @Delete(':id')
   async delete(@Param('id') id: number) {
+    if (!id) return handleBadresuest(new Error('El id es requerido'))
     return await this.clientsService.delete(id)
   }
 }
